requirements.txt
pillow
opencv-python
imgaug
ultralytics
gtts
pygame